chore(empty state): Move empty states from constants file
