+++ src/Components/PresentationalComponents/EmptyStates/EmptyStates.js
@@ -1,0 +1,176 @@
+import React from 'react';
+import propTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
+import {
+    Bullseye,
+    Button,
+    Card,
+    CardBody,
+    CardHeader,
+    EmptyStateVariant,
+    EmptyState,
+    EmptyStateBody,
+    EmptyStateIcon,
+    Text,
+    TextVariants,
+    TextContent,
+    Title
+} from '@patternfly/react-core';
+
+import { CubesIcon, ExternalLinkAltIcon, FrownOpenIcon, SecurityIcon } from '@patternfly/react-icons';
+import messages from '../../../Messages';
+
+// Reports that no CVEs were found after filtering results
+export const FilterNotFoundForCVE = (
+    <Bullseye style={{ height: 'auto' }}>
+        <EmptyState variant={EmptyStateVariant.large}>
+            <Title headingLevel="h5" size="lg">
+                <FormattedMessage {...messages.emptyStateNoMatchingCves} />
+            </Title>
+            <TextContent>
+                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
+                    <FormattedMessage {...messages.emptyStateMaybeFollowing} />
+                    <Text>
+                        <FormattedMessage {...messages.emptyStateNoCveReported} />
+                    </Text>
+                    <Text component={TextVariants.p}>
+                        <FormattedMessage {...messages.emptyStateForMoreInfo} />
+                        <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
+                            <FormattedMessage {...messages.emptyStateForMoreLink} />  <ExternalLinkAltIcon />
+                        </a>
+                    </Text>
+                </EmptyStateBody>
+            </TextContent>
+        </EmptyState>
+    </Bullseye>
+);
+
+export const FilterNotFoundForSystem = (
+    <Bullseye style={{ height: 'auto' }}>
+        <EmptyState variant={EmptyStateVariant.large}>
+            <Title headingLevel="h5" size="lg">
+                <FormattedMessage {...messages.emptyStateNoMatchingCves} />
+            </Title>
+            <TextContent>
+                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
+                    <FormattedMessage {...messages.emptyStateMaybeFollowing} />
+                    {<br></br>}
+                    <Text component={TextVariants.p}>
+                        <FormattedMessage {...messages.emptyStateNoSystemsMatching} />
+                    </Text>
+                    <Text component={TextVariants.p}>
+                        <FormattedMessage {...messages.emptyStateMaybeExcluded} />
+                    </Text>
+                    {<br></br>}
+                    <Text component={TextVariants.p}>
+                        <FormattedMessage {...messages.emptyStateIfCveForSystem} />
+                        <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
+                            <FormattedMessage {...messages.emptyStateForMoreLink} />  <ExternalLinkAltIcon />
+                        </a>
+                    </Text>
+                </EmptyStateBody>
+            </TextContent>
+        </EmptyState>
+    </Bullseye>
+);
+
+export const EmptyCVEListForSystem = (
+    <Bullseye style={{ height: 'auto' }}>
+        <EmptyState variant={EmptyStateVariant.large}>
+            <Title headingLevel="h5" size="lg">
+                <FormattedMessage {...messages.emptyStateNoReportedSystem} />
+            </Title>
+            <TextContent>
+                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
+                    <Text>
+                        <FormattedMessage {...messages.emptyStateNoCveAffects} />
+                    </Text>
+                    <Text component={TextVariants.p}>
+                        <FormattedMessage {...messages.emptyStateIfCveForSystem}/>
+                        <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
+                            <FormattedMessage {...messages.emptyStateForMoreLink} />  <ExternalLinkAltIcon />
+                        </a>.
+                    </Text>
+
+                </EmptyStateBody>
+            </TextContent>
+        </EmptyState>
+    </Bullseye>
+);
+
+export const EmptyExcludedSystem = ({ buttonAction }) => (
+    <Bullseye>
+        <EmptyState variant={EmptyStateVariant.large}>
+            <EmptyStateIcon icon={SecurityIcon} size={'sm'} />
+            <Title headingLevel="h5" size="lg">
+                <FormattedMessage {...messages.emptyStateSystemExcludedTitle} />
+            </Title>
+            <EmptyStateBody>
+                <FormattedMessage {...messages.emptyStateSystemExcludedBody} />
+            </EmptyStateBody>
+            <Button variant="primary" component="a" onClick={buttonAction}>
+                <FormattedMessage {...messages.emptyStateSystemResumeAnalysis} />
+            </Button>
+        </EmptyState>
+    </Bullseye>
+);
+
+EmptyExcludedSystem.propTypes = {
+    buttonAction: propTypes.func
+};
+
+export const EmptyCVEList = (
+    <Bullseye style={{ height: 'auto' }}>
+        <EmptyState variant={EmptyStateVariant.large}>
+            <Title headingLevel="h5" size="lg">
+                <FormattedMessage {...messages.emptyStateNoReportedSystem} />
+            </Title>
+            <TextContent>
+                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
+                    <Text>
+                        <FormattedMessage {...messages.emptyStateNoCveAffects} />
+                    </Text>
+                    <FormattedMessage {...messages.emptyStateIfCveForSystems}/>
+                    <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
+                        <FormattedMessage {...messages.emptyStateForMoreLink} /> <ExternalLinkAltIcon />
+                    </a>
+                    .
+                </EmptyStateBody>
+            </TextContent>
+        </EmptyState>
+    </Bullseye>
+);
+
+// Indicates that Insights is not activated
+export const EmptyVulnerabilityData = (
+    <Bullseye>
+        <EmptyState>
+            <EmptyStateIcon icon={CubesIcon} />
+            <Title headingLevel="h5" size="lg">
+                <FormattedMessage {...messages.emptyStateNoData} />
+            </Title>
+            <EmptyStateBody>
+                <FormattedMessage {...messages.emptyStateActivateInsights} />
+            </EmptyStateBody>
+            <Button
+                variant="primary"
+                component="a"
+                href="http://access.redhat.com/products/cloud_management_services_for_rhel#getstarted"
+            >
+                <FormattedMessage {...messages.emptyStateLearnInsights} />
+            </Button>
+        </EmptyState>
+    </Bullseye>
+);
+
+// Generic error
+export const GenericError = (
+    <Card className="ins-empty-rule-cards">
+        <CardHeader>
+            <FrownOpenIcon size="lg" />
+        </CardHeader>
+        <CardBody>
+            <FormattedMessage {...messages.errorReload} />
+        </CardBody>
+    </Card>
+);
