+++ src/Components/SmartComponents/CVEs/VulnerabilitiesCves.js
@@ -6,7 +6,8 @@ import qs from 'query-string';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import { CVSSOptions, GenericError, PublicDateOptions } from '../../../Helpers/constants';
+import { CVSSOptions, PublicDateOptions } from '../../../Helpers/constants';
+import { GenericError } from '../../PresentationalComponents/EmptyStates/EmptyStates';
 import { createCveListByAccount } from '../../../Helpers/VulnerabilitiesHelper';
 import { changeCveListParameters, fetchCveListByAccount, selectCve, expandCve } from '../../../Store/Actions/Actions';
 import BusinessRiskModal from '../Modals/BusinessRiskModal';
