+++ src/Components/SmartComponents/CVEs/VulnerabilitiesTable.js
@@ -1,11 +1,17 @@
+import React, { Component, Fragment } from 'react';
+import propTypes from 'prop-types';
 import { Table, TableBody, TableHeader } from '@patternfly/react-table';
 import { SkeletonTable, TableToolbar } from '@redhat-cloud-services/frontend-components';
-import propTypes from 'prop-types';
-import React, { Component, Fragment } from 'react';
-import { EmptyCVEList, EmptyCVEListForSystem, FilterNotFoundForCVE } from '../../../Helpers/constants';
+
 import { cveTableRowActions } from '../../../Helpers/CVEHelper';
 import { createSortBy, handleSortColumn } from '../../../Helpers/MiscHelper';
 import PaginationWrapper from '../../PresentationalComponents/PaginationWrapper/PaginationWrapper';
+import {
+    FilterNotFoundForCVE,
+    EmptyCVEListForSystem,
+    EmptyCVEList
+} from '../../PresentationalComponents/EmptyStates/EmptyStates';
+
 import { CVETableContext } from './VulnerabilitiesCves';
 
 class VulnerabilitiesTableWithContext extends Component {
@@ -22,8 +28,13 @@ class VulnerabilitiesTableWithContext extends Component {
             item => Object.prototype.hasOwnProperty.call(cves.meta, item) && cves.meta[item]
         );
 
-        let emptyComponent = (filterFields.length !== 0)
-            ? FilterNotFoundForCVE : entity ? EmptyCVEListForSystem : (cves.data.length === 0) ? EmptyCVEList : undefined;
+        let emptyComponent = filterFields.length !== 0
+            ? FilterNotFoundForCVE
+            : entity
+                ? EmptyCVEListForSystem
+                : cves.data.length === 0
+                    ? EmptyCVEList
+                    : undefined;
 
         return ([{
             heightAuto: true,
