+++ src/Components/SmartComponents/SystemCves/SystemCveTable.js
@@ -1,11 +1,16 @@
+import React, { Component, Fragment } from 'react';
+import propTypes from 'prop-types';
 import { Table, TableBody, TableHeader } from '@patternfly/react-table';
 import { SkeletonTable, TableToolbar } from '@redhat-cloud-services/frontend-components';
-import propTypes from 'prop-types';
-import React, { Component, Fragment } from 'react';
-import { EmptyCVEList, EmptyCVEListForSystem, FilterNotFoundForSystem } from '../../../Helpers/constants';
+
 import { systemCveTableRowActions } from '../../../Helpers/CVEHelper';
 import { createSortBy, handleSortColumn } from '../../../Helpers/MiscHelper';
 import PaginationWrapper from '../../PresentationalComponents/PaginationWrapper/PaginationWrapper';
+import {
+    FilterNotFoundForSystem,
+    EmptyCVEListForSystem,
+    EmptyCVEList
+} from '../../PresentationalComponents/EmptyStates/EmptyStates';
 import { CVETableContext } from './SystemCves';
 
 class SystemCvesTableWithContext extends Component {
@@ -21,8 +26,13 @@ class SystemCvesTableWithContext extends Component {
         const filterFields = ['filter', 'cvss_from', 'cvss_to', 'public_from', 'public_to', 'impact', 'status_id'].filter(
             item => Object.prototype.hasOwnProperty.call(cves.meta, item) && cves.meta[item]
         );
-        let emptyComponent = (filterFields.length !== 0)
-            ? FilterNotFoundForSystem : entity ? EmptyCVEListForSystem : (cves.data.length === 0) ? EmptyCVEList : undefined;
+        let emptyComponent = filterFields.length !== 0
+            ? FilterNotFoundForSystem
+            : entity
+                ? EmptyCVEListForSystem
+                : cves.data.length === 0
+                    ? EmptyCVEList
+                    : undefined;
 
         return ([{
             heightAuto: true,
