+++ src/Components/SmartComponents/SystemCves/SystemCves.js
@@ -1,22 +1,26 @@
 /* eslint-disable camelcase */
-import { Stack, StackItem, Text, TextContent, TextVariants } from '@patternfly/react-core';
-import { classNames, expandable, sortable } from '@patternfly/react-table';
-import propTypes from 'prop-types';
-import qs from 'query-string';
 import React, { Component, Fragment } from 'react';
+import qs from 'query-string';
+import propTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import { CVSSOptions, GenericError, NoVulnerabilityData, PublicDateOptions } from '../../../Helpers/constants';
-import { updateStateSet } from '../../../Helpers/MiscHelper';
-import { createCveListBySystem } from '../../../Helpers/VulnerabilitiesHelper';
-import { fetchCveListBySystem } from '../../../Store/Actions/Actions';
-import '../CVEs/vulnerabilities.scss';
-import CvePairStatusModal from '../Modals/CvePairStatusModal';
-import SystemCveTable from './SystemCveTable';
-import SystemCveTableToolbar from './SystemCveTableToolbar';
-import messages from '../../../Messages';
 import { injectIntl, IntlProvider } from 'react-intl';
+import { classNames, expandable, sortable } from '@patternfly/react-table';
+import { Stack, StackItem, Text, TextContent, TextVariants } from '@patternfly/react-core';
+
+import { EmptyVulnerabilityData, GenericError } from '../../PresentationalComponents/EmptyStates/EmptyStates';
+import { createCveListBySystem } from '../../../Helpers/VulnerabilitiesHelper';
+import { CVSSOptions, PublicDateOptions } from '../../../Helpers/constants';
+import { fetchCveListBySystem } from '../../../Store/Actions/Actions';
+import { updateStateSet } from '../../../Helpers/MiscHelper';
 import DownloadReport from '../../../Helpers/DownloadReport';
+import CvePairStatusModal from '../Modals/CvePairStatusModal';
+import SystemCveTableToolbar from './SystemCveTableToolbar';
+import SystemCveTable from './SystemCveTable';
+import messages from '../../../Messages';
+
+import '../CVEs/vulnerabilities.scss';
+
 export const CVETableContext = React.createContext({});
 
 class SystemCves extends Component {
@@ -49,7 +53,7 @@ class SystemCves extends Component {
         const { status } = error;
         const statusCode = parseInt(status);
         if (statusCode === 404 && this.props.entity) {
-            return NoVulnerabilityData;
+            return EmptyVulnerabilityData;
         } else {
             return GenericError;
         }
