+++ src/Components/SmartComponents/SystemDetail/SystemDetail.js
@@ -1,13 +1,14 @@
-import propTypes from 'prop-types';
 import React from 'react';
+import propTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import { SystemExludedFromAnalysis } from '../../../Helpers/constants';
-import { setHeader } from '../../../Store/Actions/Actions';
-import { paths } from '../../../Utilities/Routes';
-import { ConnectedSystemCves } from '../SystemCves/SystemCves';
-import messages from '../../../Messages';
 import { FormattedMessage } from 'react-intl';;
+
+import messages from '../../../Messages';
+import { paths } from '../../../Utilities/Routes';
+import { setHeader } from '../../../Store/Actions/Actions';
+import { ConnectedSystemCves } from '../SystemCves/SystemCves';
+import { EmptyExcludedSystem } from '../../PresentationalComponents/EmptyStates/EmptyStates';
 
 class SystemDetail extends React.Component {
     static propTypes = {
@@ -44,7 +45,7 @@ class SystemDetail extends React.Component {
         } else {
             return (
                 <React.Fragment>
-                    <SystemExludedFromAnalysis buttonAction={() => optOutSystemHandler(false)} />
+                    <EmptyExcludedSystem buttonAction={() => optOutSystemHandler(false)} />
                 </React.Fragment>
             );
         }
