+++ src/Helpers/constants.js
@@ -1,50 +1,16 @@
 /* eslint-disable no-console */
-import {
-    Bullseye,
-    Button,
-    Card,
-    CardBody,
-    CardHeader,
-    EmptyStateVariant,
-    EmptyState,
-    EmptyStateBody,
-    EmptyStateIcon,
-    Text,
-    TextVariants,
-    TextContent,
-    Title
-} from '@patternfly/react-core';
-import moment from 'moment';
-import messages from '../Messages';
 import React, { Fragment } from 'react';
-import { FormattedMessage } from 'react-intl';
-import { intl } from '../Utilities/IntlProvider';
+import moment from 'moment';
 import BellIcon from '@patternfly/react-icons/dist/js/icons/bell-icon';
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
-import { CubesIcon, ExternalLinkAltIcon, FrownOpenIcon, SecurityIcon } from '@patternfly/react-icons';
 import ExclamationTriangleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-triangle-icon';
+
+import messages from '../Messages';
+import { intl } from '../Utilities/IntlProvider';
 
 export const DEFAULT_PAGE_SIZE = 20;
 export const RH_KB_URL = 'https://access.redhat.com/node';
 export const CVES_PATH = `${document.baseURI}insights/vulnerability/cves`;
-
-// eslint-disable-next-line react/prop-types
-export const SystemExludedFromAnalysis = ({ buttonAction }) => (
-    <Bullseye>
-        <EmptyState variant={EmptyStateVariant.large}>
-            <EmptyStateIcon icon={SecurityIcon} size={'sm'} />
-            <Title headingLevel="h5" size="lg">
-                <FormattedMessage {...messages.emptyStateSystemExcludedTitle} />
-            </Title>
-            <EmptyStateBody>
-                <FormattedMessage {...messages.emptyStateSystemExcludedBody} />
-            </EmptyStateBody>
-            <Button variant="primary" component="a" onClick={buttonAction}>
-                <FormattedMessage {...messages.emptyStateSystemResumeAnalysis} />
-            </Button>
-        </EmptyState>
-    </Bullseye>
-);
 
 export const impactTextList = {
     critical: intl.formatMessage(messages.impactTextListCritical),
@@ -93,141 +59,6 @@ export const impactList = {
         text: impactTextList.low
     }
 };
-
-// Reports that no CVEs were found after filtering results
-export const FilterNotFoundForCVE = (
-    <Bullseye style={{ height: 'auto' }}>
-        <EmptyState variant={EmptyStateVariant.large}>
-            <Title headingLevel="h5" size="lg">
-                <FormattedMessage {...messages.emptyStateNoMatchingCves} />
-            </Title>
-            <TextContent>
-                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
-                    <FormattedMessage {...messages.emptyStateMaybeFollowing} />
-                    <Text>
-                        <FormattedMessage {...messages.emptyStateNoCveReported} />
-                    </Text>
-                    <Text component={TextVariants.p}>
-                        <FormattedMessage {...messages.emptyStateForMoreInfo} />
-                        <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
-                            <FormattedMessage {...messages.emptyStateForMoreLink} />  <ExternalLinkAltIcon />
-                        </a>
-                    </Text>
-                </EmptyStateBody>
-            </TextContent>
-        </EmptyState>
-    </Bullseye>
-);
-
-export const FilterNotFoundForSystem = (
-    <Bullseye style={{ height: 'auto' }}>
-        <EmptyState variant={EmptyStateVariant.large}>
-            <Title headingLevel="h5" size="lg">
-                <FormattedMessage {...messages.emptyStateNoMatchingCves} />
-            </Title>
-            <TextContent>
-                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
-                    <FormattedMessage {...messages.emptyStateMaybeFollowing} />
-                    {<br></br>}
-                    <Text component={TextVariants.p}>
-                        <FormattedMessage {...messages.emptyStateNoSystemsMatching} />
-                    </Text>
-                    <Text component={TextVariants.p}>
-                        <FormattedMessage {...messages.emptyStateMaybeExcluded} />
-                    </Text>
-                    {<br></br>}
-                    <Text component={TextVariants.p}>
-                        <FormattedMessage {...messages.emptyStateIfCveForSystem} />
-                        <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
-                            <FormattedMessage {...messages.emptyStateForMoreLink} />  <ExternalLinkAltIcon />
-                        </a>
-                    </Text>
-                </EmptyStateBody>
-            </TextContent>
-        </EmptyState>
-    </Bullseye>
-);
-// Selected system do not have any CVEs
-export const EmptyCVEListForSystem = (
-    <Bullseye style={{ height: 'auto' }}>
-        <EmptyState variant={EmptyStateVariant.large}>
-            <Title headingLevel="h5" size="lg">
-                <FormattedMessage {...messages.emptyStateNoReportedSystem} />
-            </Title>
-            <TextContent>
-                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
-                    <Text>
-                        <FormattedMessage {...messages.emptyStateNoCveAffects} />
-                    </Text>
-                    <Text component={TextVariants.p}>
-                        <FormattedMessage {...messages.emptyStateIfCveForSystem}/>
-                        <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
-                            <FormattedMessage {...messages.emptyStateForMoreLink} />  <ExternalLinkAltIcon />
-                        </a>.
-                    </Text>
-
-                </EmptyStateBody>
-            </TextContent>
-        </EmptyState>
-    </Bullseye>
-);
-
-// None of my systems is affected
-export const EmptyCVEList = (
-    <Bullseye style={{ height: 'auto' }}>
-        <EmptyState variant={EmptyStateVariant.large}>
-            <Title headingLevel="h5" size="lg">
-                <FormattedMessage {...messages.emptyStateNoReportedSystem} />
-            </Title>
-            <TextContent>
-                <EmptyStateBody style={{ whiteSpace: 'normal' }}>
-                    <Text>
-                        <FormattedMessage {...messages.emptyStateNoCveAffects} />
-                    </Text>
-                    <FormattedMessage {...messages.emptyStateIfCveForSystems}/>
-                    <a href="https://access.redhat.com/security/team/contact/" target="__blank" rel="noopener noreferrer">
-                        <FormattedMessage {...messages.emptyStateForMoreLink} /> <ExternalLinkAltIcon />
-                    </a>
-                    .
-                </EmptyStateBody>
-            </TextContent>
-        </EmptyState>
-    </Bullseye>
-);
-
-// Indicates that Insights is not activated
-export const NoVulnerabilityData = (
-    <Bullseye>
-        <EmptyState>
-            <EmptyStateIcon icon={CubesIcon} />
-            <Title headingLevel="h5" size="lg">
-                <FormattedMessage {...messages.emptyStateNoData} />
-            </Title>
-            <EmptyStateBody>
-                <FormattedMessage {...messages.emptyStateActivateInsights} />
-            </EmptyStateBody>
-            <Button
-                variant="primary"
-                component="a"
-                href="http://access.redhat.com/products/cloud_management_services_for_rhel#getstarted"
-            >
-                <FormattedMessage {...messages.emptyStateLearnInsights} />
-            </Button>
-        </EmptyState>
-    </Bullseye>
-);
-
-// Generic error
-export const GenericError = (
-    <Card className="ins-empty-rule-cards">
-        <CardHeader>
-            <FrownOpenIcon size="lg" />
-        </CardHeader>
-        <CardBody>
-            <FormattedMessage {...messages.errorReload} />
-        </CardBody>
-    </Card>
-);
 
 // Read-only mode notification text
 export const ReadOnlyNotification = {
